Move ContentUtil to the internal package
