+++ langchain4j-core/src/main/java/dev/langchain4j/data/message/ImageContent.java
@@ -7,7 +7,7 @@ import java.nio.file.Path;
 import java.util.Objects;
 
 import static dev.langchain4j.data.message.ContentType.IMAGE;
-import static dev.langchain4j.data.message.ContentUtil.extractBase64Content;
+import static dev.langchain4j.internal.ContentUtil.extractBase64Content;
 import static dev.langchain4j.data.message.ImageContent.DetailLevel.LOW;
 import static dev.langchain4j.internal.ValidationUtils.ensureNotBlank;
 import static dev.langchain4j.internal.ValidationUtils.ensureNotNull;
